TooManyItems 1.0

+++ private/controllers/Popup.php
@@ -4,6 +4,10 @@
             
         }
         public function marca(){
+            if (!($_SESSION['permisos']['admin'] || $_SESSION['permisos']['coord'])) {
+                header('location:' . URLROOT . '/ErrorController/permisos');
+                die();
+            }
             require_once APPROOT. '/models/Marca.php';
             $marcaModel = new Marca();
             $marca=[
@@ -12,7 +16,21 @@
             $marcaModel->insertMarca($marca, $_POST['sector']);
             header('location:' . URLROOT . $_POST['origen']);
         }
+        public function eliminarMarca($codMarca){
+            if (!($_SESSION['permisos']['admin'] || $_SESSION['permisos']['coord'])) {
+                header('location:' . URLROOT . '/ErrorController/permisos');
+                die();
+            }
+            require_once APPROOT . '/models/Marca.php';
+            $marcaModel = new Marca();
+            $marcaModel->deleteMarca($codMarca);
+            //header('location:' . URLROOT . $_POST['origen']);
+        }
         public function proveedor(){
+            if (!($_SESSION['permisos']['admin'] || $_SESSION['permisos']['coord'])) {
+                header('location:' . URLROOT . '/ErrorController/permisos');
+                die();
+            }
             require_once APPROOT . '/models/Instancia/Proveedor.php';
             $proveedorModel = new Proveedor();
             $proveedor = [
@@ -22,7 +40,21 @@
             $proveedorModel->insertProveedor($proveedor, $_POST['sector']);
             header('location:' . URLROOT . $_POST['origen']);
         }
+        public function eliminarProveedor($codProveedor){
+            if (!($_SESSION['permisos']['admin'] || $_SESSION['permisos']['coord'])) {
+                header('location:' . URLROOT . '/ErrorController/permisos');
+                die();
+            }
+            require_once APPROOT . '/models/Instancia/Proveedor.php';
+            $marcaModel = new Proveedor();
+            $marcaModel->deleteProveedor($codProveedor);
+            header('location:' . URLROOT . $_POST['origen']);
+        }
         public function ubicacion(){
+            if (!($_SESSION['permisos']['admin'] || $_SESSION['permisos']['coord'])) {
+                header('location:' . URLROOT . '/ErrorController/permisos');
+                die();
+            }
             require_once APPROOT . '/models/Instancia/Ubicacion.php';
             $ubicacionModel = new Ubicacion();
             $ubicacion = [
@@ -32,19 +64,41 @@
             $ubicacionModel->insertUbicacion($ubicacion, $_POST['sector']);
             header('location:' . URLROOT . $_POST['origen']);
         }
+        public function eliminarUbicacion($codUbicacion){
+            if (!($_SESSION['permisos']['admin'] || $_SESSION['permisos']['coord'])) {
+                header('location:' . URLROOT . '/ErrorController/permisos');
+                die();
+            }
+            require_once APPROOT . '/models/Instancia/Ubicacion.php';
+            $marcaModel = new Ubicacion();
+            $marcaModel->deleteUbicacion($codUbicacion);
+            header('location:' . URLROOT . $_POST['origen']);
+        }
         public function eliminarInsumo($codInsumo, $codSector){
+            if (!($_SESSION['permisos']['admin'] || $_SESSION['permisos']['coord'])) {
+                header('location:' . URLROOT . '/ErrorController/permisos');
+                die();
+            }
             require_once APPROOT . '/models/Insumo.php';
             $insumoModel= new Insumo();
             $insumoModel->deleteInsumo($codInsumo,$codSector);
             header('location:' . URLROOT . $_POST['origen']);
         }
         public function eliminarInstancia($codInstancia, $codSector, $codInsumo){
+            if (!($_SESSION['permisos']['admin'] || $_SESSION['permisos']['coord'])) {
+                header('location:' . URLROOT . '/ErrorController/permisos');
+                die();
+            }
             require_once APPROOT . '/models/Instancia.php';
             $instanciaModel= new Instancia();
             $instanciaModel->deleteInstancia($codInstancia, $codSector, $codInsumo);
             header('location:' . URLROOT . $_POST['origen']);
         }
         public function eliminarCompra($codCompra, $codInsumo, $codSector){
+            if (!($_SESSION['permisos']['admin'] || $_SESSION['permisos']['coord'])) {
+                header('location:' . URLROOT . '/ErrorController/permisos');
+                die();
+            }
             require_once APPROOT . '/models/Instancia.php';
             $instanciaModel = new Instancia();
             $instanciaModel->deleteCompra($codCompra, $codInsumo, $codSector);
@@ -69,6 +123,10 @@
             header('location:' . URLROOT . $_POST['origen']);
         }
         public function agregarPrestamo(){
+            if (!($_SESSION['permisos']['admin'] || $_SESSION['permisos']['panio'] || $_SESSION['permisos']['docente'])) {
+                header('location:' . URLROOT . '/ErrorController/permisos');
+                die();
+            }
             $i=1;
             $insumosSeleccionados=[];
             while (isset($_POST['insumo'.$i])) {
@@ -90,6 +148,7 @@
             $ci=isset($_POST['nombreAlumno']) ? $_POST['nombreAlumno'] : $_SESSION['cuenta']['ci'];
             $prestamo=[
                 'clase' => $_POST['claseAlumno'],
+                'codSector' => $_POST['sector'],
                 'ciPrestatario' => $ci,
                 'fecha' => $_POST['fechaPrestamo'],
                 'hora' => $_POST['horaPrestamo'],
@@ -104,9 +163,19 @@
             header('location:' . URLROOT . $_POST['origen']);
         }
         public function devolverPrestamo($codPrestamo){
+            if (!($_SESSION['permisos']['admin'] || $_SESSION['permisos']['panio'] || $_SESSION['permisos']['docente'])) {
+                header('location:' . URLROOT . '/ErrorController/permisos');
+                die();
+            }
             require_once APPROOT . '/models/Prestamo.php';
             $prestamoModel = new PrestamoModel();
             $prestamoModel->devolverPrestamo($codPrestamo);
+            header('location:' . URLROOT . $_POST['origen']);
+        }
+        public function cambiarContrasenia(){
+            require_once APPROOT . '/models/Cuenta.php';
+            $prestamoModel = new Usuario();
+            $prestamoModel->cambiarContrasenia($_POST['contrasenia']);
             header('location:' . URLROOT . $_POST['origen']);
         }
     }
